+++ src/app/feature/cart/cart.module.ts
@@ -3,10 +3,11 @@ import { CommonModule } from '@angular/common';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { CartService } from '../shared/services/cart.service';
 import { CartListComponent } from './components/cart-list/cart-list.component';
-
+import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [CommonModule,
+    SharedModule],
   declarations: [
     CartItemComponent,
     CartListComponent],