+++ src/app/feature/shared/shared.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CoolDirective } from './background/cool.directive';
+
+@NgModule({
+  declarations: [
+    CoolDirective
+  ],
+  exports:[
+    CoolDirective
+  ],
+})
+export class SharedModule { }